backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
redis
celery
gunicorn
python-dotenv
Flask-Migrate
Flask-Cors
